requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
ollama==0.1.7
sqlalchemy==2.0.23
alembic==1.13.0
fastapi==0.109.0
uvicorn==0.25.0

websockets==12.0
loguru==0.7.2
rich==13.7.0
tqdm==4.66.1
Pillow==10.2.0
pytest==7.4.4
pytest-asyncio==0.23.3